# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NexaTalent Enterprise Recruitment API",
    "version": "2.0.0",
    "description": "All 5 spec phases. Real JWT (15m access / 7d refresh rotation), bcrypt, 20 roles, 17 permissions, file-persisted DB, zod validation, idempotent finance, interview-gated chat. Legacy paths kept for old frontend builds."
  },
  "servers": [{ "url": "http://localhost:5000" }],
  "security": [{ "bearerAuth": [] }],
  "components": { "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } } },
  "paths": {
...
    "/api/v1/auth/login": { "post": { "summary": "Phase 0/1: bcrypt login, role-locked 403", "responses": { "200": { "description": "JWT pair" }, "401": { "description": "bad credentials" }, "403": { "description": "suspended or wrong workspace" } } } },
    "/api/v1/auth/mfa/setup": { "post": { "summary": "Start TOTP enrol (returns secret + otpauth URL)", "responses": { "200": { "description": "ok" }, "409": { "description": "already enabled" } } } },
    "/api/v1/auth/mfa/verify": { "post": { "summary": "Confirm TOTP code, enable MFA, one-time backup codes", "responses": { "200": { "description": "ok" }, "401": { "description": "bad code" } } } },
    "/api/v1/auth/mfa/challenge": { "post": { "summary": "Complete MFA login with ticket + otp|backupCode", "responses": { "200": { "description": "tokens" }, "401": { "description": "bad/expired" } } } },
...
    "/api/v1/auth/mfa/reset": { "post": { "summary": "Admin MFA reset (perm: manage_permissions, audited)", "responses": { "200": { "description": "ok" } } } },    "/api/v1/auth/refresh": { "post": { "summary": "Rotate JWT pair, reuse detection", "responses": { "200": { "description": "new pair" }, "401": { "description": "reused/invalid" } } } },
...
  }
}